msm: kgsl: Change gpuaddr to unsigned long everywhere
gpuaddr was changed to unsigned long in certain structs but not
everywhere. Also fix some print statements to regard gpuaddr as
type long.

+++ include/linux/msm_kgsl.h
@@ -201,7 +201,7 @@ enum kgsl_property_type {
 };
 
 struct kgsl_shadowprop {
-	unsigned int gpuaddr;
+	unsigned long gpuaddr;
 	unsigned int size;
 	unsigned int flags; /* contains KGSL_FLAGS_ values */
 };
@@ -236,7 +236,7 @@ struct kgsl_version {
 
 /* structure holds list of ibs */
 struct kgsl_ibdesc {
-	unsigned int gpuaddr;
+	unsigned long gpuaddr;
 	void *hostptr;
 	unsigned int sizedwords;
 	unsigned int ctrl;
@@ -330,7 +330,7 @@ struct kgsl_cmdstream_readtimestamp {
  * type should be a value from enum kgsl_timestamp_type
  */
 struct kgsl_cmdstream_freememontimestamp {
-	unsigned int gpuaddr;
+	unsigned long gpuaddr;
 	unsigned int type;
 	unsigned int timestamp;
 };
@@ -370,10 +370,10 @@ struct kgsl_drawctxt_destroy {
  * into the GPU address space */
 struct kgsl_map_user_mem {
 	int fd;
-	unsigned int gpuaddr;   /*output param */
+	unsigned long gpuaddr;   /*output param */
 	unsigned int len;
 	unsigned int offset;
-	unsigned int hostptr;   /*input param */
+	unsigned long hostptr;   /*input param */
 	enum kgsl_user_mem_type memtype;
 	unsigned int flags;
 };
@@ -392,7 +392,7 @@ struct kgsl_cmdstream_readtimestamp_ctxtid {
 
 struct kgsl_cmdstream_freememontimestamp_ctxtid {
 	unsigned int context_id;
-	unsigned int gpuaddr;
+	unsigned long gpuaddr;
 	unsigned int type;
 	unsigned int timestamp;
 };
@@ -404,7 +404,7 @@ struct kgsl_cmdstream_freememontimestamp_ctxtid {
 /* add a block of pmem or fb into the GPU address space */
 struct kgsl_sharedmem_from_pmem {
 	int pmem_fd;
-	unsigned int gpuaddr;	/*output param */
+	unsigned long gpuaddr;	/*output param */
 	unsigned int len;
 	unsigned int offset;
 };
@@ -414,7 +414,7 @@ struct kgsl_sharedmem_from_pmem {
 
 /* remove memory from the GPU's address space */
 struct kgsl_sharedmem_free {
-	unsigned int gpuaddr;
+	unsigned long gpuaddr;
 };
 
 #define IOCTL_KGSL_SHAREDMEM_FREE \
@@ -470,7 +470,7 @@ struct kgsl_bind_gmem_shadow {
  */
 
 struct kgsl_sharedmem_from_vmalloc {
-	unsigned int gpuaddr;	/*output param */
+	unsigned long gpuaddr;	/*output param */
 	unsigned int hostptr;
 	unsigned int flags;
 };
@@ -526,7 +526,7 @@ struct kgsl_gpumem_alloc {
 	_IOWR(KGSL_IOC_TYPE, 0x2f, struct kgsl_gpumem_alloc)
 
 struct kgsl_cff_syncmem {
-	unsigned int gpuaddr;
+	unsigned long gpuaddr;
 	unsigned int len;
 	unsigned int __pad[2]; /* For future binary compatibility */
 };
@@ -668,7 +668,7 @@ struct kgsl_gpumem_get_info {
  *
  */
 struct kgsl_gpumem_sync_cache {
-	unsigned int gpuaddr;
+	unsigned long gpuaddr;
 	unsigned int id;
 	unsigned int op;
 /* private: reserved for future use*/
